+++ routes/slots.js
@@ -1,10 +1,7 @@
+const BookedSlots = require("../models/BookedSlots");
 const Slots = require("../models/Slots");
 // const Places = require("../models/Slots");
-const {
-  verifyToken,
-  verifyTokenAndAuthorization,
-  verifyTokenAndAdmin,
-} = require("./verifyToken");
+const { verifyTokenAndAdmin } = require("./verifyToken");
 const router = require("express").Router();
 
 // CREATE
@@ -38,32 +35,46 @@ router.put("/", verifyTokenAndAdmin, async (req, res) => {
 //Delete
 router.delete("/", verifyTokenAndAdmin, async (req, res) => {
   try {
-    await Slots.findByIdAndDelete(req.body.SlotID);
-    res.status(200).json("Place deleted successfully.....");
+    const BookingToBeDeleted = await BookedSlots.find({
+      parkingPlaceID: {
+        $in: req.body.parkingPlaceID,
+      },
+      slotNo: {
+        $in: req.body.slotNo,
+      },
+    });
+
+    const d = new Date().getTime();
+    const BooleanArray = BookingToBeDeleted.map((item) => {
+      return item.from >= d || item.to >= d ? true : false;
+    });
+
+    const state = BooleanArray.find((item) => {
+      return item === true;
+    });
+    let response;
+    if (state) {
+      response = "Unable to delete";
+    } else {
+      BookingToBeDeleted.forEach(async (item) => {
+        await BookedSlots.findByIdAndDelete(item._id.toString());
+      });
+
+      await Slots.findByIdAndDelete(req.body.SlotID);
+      response = "Deleted successfully";
+    }
+    return res.status(200).json(response);
   } catch (err) {
     res.status(500).json(err);
   }
 });
 
-//Get all PlaceS
+//Get all slots
 
 router.get("/", verifyTokenAndAdmin, async (req, res) => {
-  const qNew = req.query.new;
-  const qCategory = req.query.category;
   try {
-    let GetPlaces;
-    if (qNew) {
-      GetPlaces = await Places.find().sort({ createdAt: -1 }).limit(1);
-    } else if (qCategory) {
-      GetPlaces = await Places.find({
-        categories: {
-          $in: [qCategory],
-        },
-      });
-    } else {
-      GetSlots = await Slots.find();
-    }
-    res.status(200).json(GetSlots);
+    GetAllSlots = await Slots.find();
+    res.status(200).json(AllGetSlots);
   } catch (err) {
     res.status(500).json(err);
   }
